ref(api): split validators into per domain modules
